Fix la configuration tailwind
