+++ apps/app/tsconfig.app.json
@@ -5,9 +5,6 @@
   },
   "references": [
     {
-      "path": "../../packages/pdf-components/tsconfig.lib.json"
-    },
-    {
       "path": "../../packages/ui/tsconfig.lib.json"
     },
     {
