+++ packages/pdf-components/src/primitives/Badge/BadgeBudget.tsx
@@ -1,14 +1,14 @@
 import { Text } from '@react-pdf/renderer';
+import { SizeVariant } from '@tet/design-tokens';
 import classNames from 'classnames';
 import { getFormattedNumber } from '../../fiche-action/external-helpers';
-import { BadgeSize } from '../../ui-compat';
 import { tw } from '../utils';
 import { Badge } from './Badge';
 
 type BadgeBudgetProps = {
   montant: number | undefined | null;
   unite?: 'HT' | 'ETP';
-  size?: BadgeSize;
+  size?: SizeVariant;
   className?: string;
 };
