+++ packages/ui/src/design-system/Badge/index.ts
@@ -1,4 +1,3 @@
 export * from './Badge';
 export * from './BadgeDouble';
 export * from './DEPRECATED_BadgesContainer';
-export * from './utils';
