--- packages/ui/src/design-system/Badge/utils.ts
@@ -1,1 +1,0 @@
-export { badgeClassnames } from '@tet/design-tokens';
